driver: spi_flash_w25qxxdv: no data expected on writing register byte
The functions using `spi_flash_wb_reg_write` do not expect any feedback.
Even worse, for SPIM_NRF5 SPI legacy driver, this causes two bytes to be
exchanged with the slave. The second (dummy) byte confuses this flash on
NRF5 platforms so that the Write Enable flag is never set.

Signed-off-by: Stefan Smarzly <stefan.smarzly@tado.com>